Retrieve and update the PEP 621 project version, if possible
When determining the current version, and if
`tool.bumpversion.current_version` is not set, attempt to retrieve the
version from `project.version` à la PEP 621. If that setting is not
set, or if the version is explicitly marked as dynamically set, then
continue with querying SCM tags.

When updating the configuration during bumping, if we previously
successfully retrieved a PEP 621 version, then update the `project.version`
field in `pyproject.toml` as well.