Added devenv.nix and justfile
Addition of devenv.nix allows developers to easily create hermetic
environments in order to develop this package. This greatly simplifies
the setup of the environment, and utilizes uv's tooling to create the
virtual environment(s). Devenv.nix can also handle git-hooks, but that
is already handled via the .pre-commit-config.yaml file.

The justfile includes some helpful starter recipes. Additional recipes
can be added, such as build and publishing of the package/Docker.